Added widgetmanager class
