+++ docs/_source/add_widgets_manager.rst
@@ -1,0 +1,15 @@
+
+:orphan:
+
+.. This page is orphan because its content concerns the internal
+.. working of the library. However it is necessary in order to
+.. be able to quote its items in the documentation.
+
+.. module:: pygame_menu._widgetmanager
+
+=============
+WidgetManager
+=============
+
+.. autoclass:: pygame_menu._widgetmanager.WidgetManager
+    :members:
