+++ pygame_menu/baseimage.py
@@ -62,7 +62,7 @@ from pathlib import Path
 import pygame
 import pygame_menu.locals as _locals
 from pygame_menu.utils import assert_vector, assert_position
-from pygame_menu._custom_types import Tuple2IntType, Union, Vector2NumberType, Callable, Tuple, List, \
+from pygame_menu._types import Tuple2IntType, Union, Vector2NumberType, Callable, Tuple, List, \
     NumberType, Optional, Dict, Tuple4IntType, Literal, Tuple2NumberType, ColorType, Tuple3IntType, Any
 
 # Example image paths
