+++ pygame_menu/scrollarea.py
@@ -34,11 +34,11 @@ __all__ = ['ScrollArea', 'get_scrollbars_from_position']
 import pygame
 import pygame_menu.baseimage as _baseimage
 import pygame_menu.locals as _locals
-from pygame_menu.decorator import Decorator
+from pygame_menu._decorator import Decorator
 from pygame_menu.utils import make_surface, assert_color, assert_position
 from pygame_menu.widgets import ScrollBar, MenuBar
 
-from pygame_menu._custom_types import ColorType, Union, NumberType, Tuple, List, Dict, \
+from pygame_menu._types import ColorType, Union, NumberType, Tuple, List, Dict, \
     TYPE_CHECKING, Tuple2NumberType, Optional, Tuple2IntType
 
 if TYPE_CHECKING:
