+++ pygame_menu/widgets/selection/highlight.py
@@ -33,7 +33,7 @@ __all__ = ['HighlightSelection']
 
 import pygame
 from pygame_menu.widgets.core import Selection
-from pygame_menu._custom_types import NumberType, TYPE_CHECKING
+from pygame_menu._types import NumberType, TYPE_CHECKING
 
 if TYPE_CHECKING:
     from pygame_menu.widgets.core import Widget
