+++ pygame_menu/widgets/widget/button.py
@@ -35,7 +35,7 @@ import pygame
 from pygame_menu.utils import is_callable
 from pygame_menu.widgets.core import Widget
 import pygame_menu.controls as _controls
-from pygame_menu._custom_types import Any, CallbackType, Callable, TYPE_CHECKING, Union, List, Tuple, Optional
+from pygame_menu._types import Any, CallbackType, Callable, TYPE_CHECKING, Union, List, Tuple, Optional
 
 if TYPE_CHECKING:
     from pygame_menu.menu import Menu
