+++ pygame_menu/widgets/widget/image.py
@@ -37,7 +37,7 @@ from pathlib import Path
 import pygame
 from pygame_menu.baseimage import BaseImage
 from pygame_menu.widgets import Widget
-from pygame_menu._custom_types import Union, List, NumberType, CallbackType, Tuple2NumberType, Tuple, Optional
+from pygame_menu._types import Union, List, NumberType, CallbackType, Tuple2NumberType, Tuple, Optional
 from pygame_menu.utils import assert_vector
