+++ pygame_menu/widgets/widget/textinput.py
@@ -38,7 +38,7 @@ import pygame_menu.controls as _controls
 import pygame_menu.locals as _locals
 from pygame_menu.utils import check_key_pressed_valid, make_surface, assert_color
 from pygame_menu.widgets.core import Widget
-from pygame_menu._custom_types import Optional, Any, CallbackType, Union, Tuple, List, ColorType, \
+from pygame_menu._types import Optional, Any, CallbackType, Union, Tuple, List, ColorType, \
     NumberType, Tuple2IntType, Dict, Tuple2NumberType, Literal
 
 try:
